+++ src/lib.rs
@@ -806,11 +806,11 @@ macro_rules! hopper_dynamic_fields {
 // the top-level `hopper::*` path without needing to reach through
 // `hopper_runtime::`.
 pub use hopper_runtime::{
-    address, declare_id, default_allocator, err, error, fast_entrypoint, hopper_emit_cpi,
-    hopper_entrypoint, hopper_exact_entrypoint, hopper_fast_entrypoint, hopper_lazy_entrypoint,
-    hopper_log, hopper_unsafe_region, lazy_entrypoint, msg, no_allocator, nostd_panic_handler,
-    program_entrypoint, require, require_eq, require_gt, require_gte, require_keys_eq,
-    require_keys_neq, require_lt, require_lte, require_neq,
+    address, const_pda, declare_id, default_allocator, err, error, fast_entrypoint,
+    hopper_emit_cpi, hopper_entrypoint, hopper_exact_entrypoint, hopper_fast_entrypoint,
+    hopper_lazy_entrypoint, hopper_log, hopper_unsafe_region, lazy_entrypoint, msg, no_allocator,
+    nostd_panic_handler, program_entrypoint, require, require_eq, require_gt, require_gte,
+    require_keys_eq, require_keys_neq, require_lt, require_lte, require_neq,
 };
 
 /// Declare a bounded multi-layout interface account resolver.